Remove unused dead code clusters
Delete orphaned landing-page components (DidYouKnow, SiteFeatures,
ScrollAnimationSection, etc.), superseded components (TopList,
LearnMoreCard, MunicipalityDetailSEO), unused hooks (useNationalData,
useLandingPageSEO, usePopularHeroItems), and stale files (App.css,
utils/sorting.ts, utils/detail/seo.ts).

Move HeroSearchResult type to src/types/landing.ts and remove the
unused getNationalData API stub.
