+++ src/components/landing/LandingPageCTA.tsx
@@ -6,7 +6,7 @@ import { Input } from "@/components/ui/input";
 import { LandingSection } from "./LandingSection";
 import { useLanguage } from "@/components/LanguageProvider";
 import { getEntityDetailPath, localizedPath } from "@/utils/routing";
-import type { HeroSearchResult } from "@/hooks/usePopularHeroItems";
+import type { HeroSearchResult } from "@/types/landing";
 import { Text } from "../ui/text";
 import { POPULAR_HERO_ITEMS } from "@/lib/constants/landingPage";
 import { useHeroGlobalSearch } from "../../hooks/landing/useHeroGlobalSearch";
